Fix nasty bug with PlaybackHandle
In debug mode, a nullptr will probably not evaluate to nullptr if it's from uninitialized memory. So in the handle, I was doing if (m_engine != nullptr). But it was something like 0xcdcdcd, so it went in and did stuff. That stuff was very bad, and a mutex deep inside was like "hey I'm not even initialized". So I explicitely initialy m_engine to nullptr now