Removed synchronization from ParameterizedEffect
It was originally there because I was trying to solve a sync bug. I am not sure if it actually resolved it. I hope the cause was unrelated. I am removing sync because I want to copy the events, and std::mutex is not copyable, and I don't want to write the bulletproof copy/move constructors.